Fixed texture binding and drawing function decl.
-Fixed texture binding. This fixed the bug where drawing a circle before
drawing a text would make the circle invisible. That was because texture
was bound to 1 while enigma::bound_texture was 0. This broke
texture_use() and texture_reset() which made draw_circle() (and many
other functions which use texture_reset()) to draw a circle with a bound
texture which made it invisible.
-Fixed some types so it actually compiles. Robert will probably break
them again with his scalar changes. We really need to set up those tests
so bugs like these could be noticed before commiting.

+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GSd3d.h
@@ -16,7 +16,7 @@
 **/
 
 #ifndef ENIGMA_GL3D3D_H
-#define ENIGMA_GL3D3D_H 
+#define ENIGMA_GL3D3D_H
 
 #include <string>
 
@@ -39,7 +39,7 @@ enum {
   rs_fastest,
   rs_dontcare,
   rs_exp,
-  rs_exp2, 
+  rs_exp2,
   rs_linear,
   rs_never, 	// Always False
   rs_less, 	// source Z < depth Z
@@ -142,7 +142,7 @@ bool d3d_transform_stack_disgard();
 // ***** TRANSFORMATIONS END *****
 
 // ***** LIGHTS BEGIN *****
-bool d3d_light_define_direction(int id, double dx, double dy, double dz, int col);
+bool d3d_light_define_direction(int id, float dx, float dy, float dz, int col);
 bool d3d_light_define_point(int id, float x, float y, float z, double range, int col);
 bool d3d_light_define_specularity(int id, int r, int g, int b, double a);
 void d3d_light_specularity(int facemode, int r, int g, int b, double a);
@@ -153,4 +153,4 @@ bool d3d_light_enable(int id, bool enable);
 
 }
 
-#endif // ENIGMA_GL3D3D_H 
+#endif // ENIGMA_GL3D3D_H
